requirements.txt
# Core dependencies
python-deriv-api==0.1.4
websocket-client==1.6.4
requests==2.31.0

# Data analysis
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
ta-lib==0.4.28

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.2
lightgbm==4.1.0

# Technical Analysis
pandas-ta==0.3.14b0

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
colorama==0.4.6
tabulate==0.9.0

# Performance
joblib==1.3.2
numba==0.58.1

# Logging
loguru==0.7.2
